replace int with bigInteger and decimal with bigDecimal ... currently not compiling - conversion not finished

+++ src/commonMain/kotlin/lupos/s04arithmetikOperators/multiinput/AOPMultiplication.kt
@@ -31,7 +31,7 @@ class AOPMultiplication(query: Query, childA: AOPBase, childB: AOPBase) : AOPBin
                 } else if (a is ValueFloat || b is ValueFloat) {
                     res = ValueFloat(a.toDouble() * b.toDouble())
                 } else if (a is ValueDecimal || b is ValueDecimal) {
-                    res = ValueDecimal(a.toDouble() * b.toDouble())
+                    res = ValueDecimal(a.toDecimal() * b.toDecimal())
                 } else if (a is ValueInteger || b is ValueInteger) {
                     res = ValueInteger(a.toInt() * b.toInt())
                 }
